pyproject.toml
[project]
name = "be-aura-autonomous-recommendation-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.13.0",
    "python-dotenv>=1.1.1",
]
